requirements.txt
streamlit
pandas
sqlalchemy
mysql-connector-python
python-dotenv
plotly
